PHPC-2220: Return int64 values for server connection ID (#1418)
* Update to latest libmongoc version

* PHPC-2220: Return int64 values for server connection ID

Due to BC constraints, we can't return 64-bit values on 32-bit platforms. On those platforms, the value will be truncated and an E_WARNING is raised to inform the user.

+++ src/phongo_structs.h
@@ -291,7 +291,7 @@ typedef struct {
 	zval        z_error;
 	bool        has_service_id;
 	bson_oid_t  service_id;
-	int32_t     server_connection_id;
+	int64_t     server_connection_id;
 	zend_object std;
 } php_phongo_commandfailedevent_t;
 
@@ -305,7 +305,7 @@ typedef struct {
 	char*       database_name;
 	bool        has_service_id;
 	bson_oid_t  service_id;
-	int32_t     server_connection_id;
+	int64_t     server_connection_id;
 	zend_object std;
 } php_phongo_commandstartedevent_t;
 
@@ -319,7 +319,7 @@ typedef struct {
 	bson_t*     reply;
 	bool        has_service_id;
 	bson_oid_t  service_id;
-	int32_t     server_connection_id;
+	int64_t     server_connection_id;
 	zend_object std;
 } php_phongo_commandsucceededevent_t;
